feat(reports): create reports/coverage module and move JaCoCo parser
Create dedicated reports/coverage module to consolidate coverage-related
code and prepare for adding support for additional coverage formats
(Cobertura, LCOV).

## Changes

### Module Structure
- Create `reports/coverage` submodule with standard Gradle/Kotlin layout
- Add `reports/coverage` to settings.gradle.kts
- Configure build.gradle.kts with `api(project(":reports:api"))` dependency

### Coverage Model
- Move coverage DTOs to `reports/coverage/model/CoverageModel.kt`:
  - BranchInfo
  - LineCoverage
  - FileCoverageSummary
  - FileCoverageData
  - CoverageSummary
  - CoverageResponse
- Keep `GetCoverageParams` in groovy-lsp (LSP-specific)

### Parser Abstraction
- Create `CoverageParser` interface with methods:
  - `parseWorkspace(workspaceRoot: File): CoverageResponse`
  - `parseReportFile(file: File, workspaceRoot: File): List<FileCoverageData>`
- Rename and move `JacocoXmlParser` to `JacocoParser`
- Implement `CoverageParser` interface in `JacocoParser`

### Dependencies
- Update groovy-lsp to depend on `reports:coverage` instead of `reports:api`
- groovy-lsp now transitively includes reports:api through reports:coverage

### Tests
- Move JacocoXmlParserTest to reports/coverage as JacocoParserTest
- Update test imports and package references
- Copy test resources (jacoco.xml)
- All tests pass

## Motivation

Consolidates coverage-related code into a dedicated module, following the
same pattern as reports/api. Prepares for adding Cobertura and LCOV parser
support in future PRs.
